batch_triage/requirements.txt
httpx>=0.24.0
aiometer>=0.5.0
aiosqlite>=0.19.0
jmespath>=1.0.0
tenacity>=8.2.0
tqdm>=4.65.0
pandas>=2.0.0

# Optional dependencies for export formats
openpyxl>=3.1.0    # For Excel export (.xlsx)
pyarrow>=12.0.0    # For Parquet export (.parquet)

# Parser dependencies
py-mini-racer>=0.6.0    # For JavaScript parser support

PyYAML>=6.0.1
wandb>=0.16.0
openai>=1.40.0

# Token estimation (optional but recommended for cost accuracy)
tiktoken>=0.7.0

